+++ server/component/user/UserModel.php
@@ -319,16 +319,8 @@ class UserModel extends BaseModel
                 ));
             }
             $res &= $this->db->remove_by_fk('scheduledJobs_users', 'id_users', $uid);
-
             //remove external data
-            foreach ($this->db->select_table('uploadTables') as $row => $table) {
-                foreach ($this->user_input->get_data($table['id'], '', true, FORM_EXTERNAL, $uid) as $key => $row_record) {
-                    $res &= $this->db->remove_by_ids("uploadRows", array(
-                        "id" => $row_record['record_id'],
-                        "id_uploadTables" => $table['id']
-                    ));
-                }
-            }
+            $res &= $this->db->remove_by_fk('uploadRows', 'id_users', $uid);
             $this->db->commit();
             return $res;
         } catch (Exception $e) {