+++ server/db/FUN_PRO_VIEWS/01_add_index.sql
@@ -1,0 +1,23 @@
+DELIMITER //
+DROP PROCEDURE IF EXISTS add_index //
+CREATE PROCEDURE add_index(param_table VARCHAR(100), param_index_name VARCHAR(100), param_index_column VARCHAR(1000))
+BEGIN	
+    SET @sqlstmt = (SELECT IF(
+		(
+			SELECT COUNT(*)
+            FROM information_schema.STATISTICS 
+			WHERE `table_schema` = DATABASE()
+			AND `table_name` = param_table
+            AND `index_name` = param_index_name
+		) > 0,
+        "SELECT 'The index already exists in the table'",
+        CONCAT('CREATE INDEX ', param_index_name, ' ON ', param_table, ' (', param_index_column, ');')
+    ));
+	PREPARE st FROM @sqlstmt;
+	EXECUTE st;
+	DEALLOCATE PREPARE st;	
+END
+
+//
+
+DELIMITER ;