+++ server/db/FUN_PRO_VIEWS/15_proc_get_uploadTable_with_filter.sql
@@ -2,7 +2,8 @@ DELIMITER //
 
 DROP PROCEDURE IF EXISTS get_uploadTable_with_filter //
 
-CREATE PROCEDURE get_uploadTable_with_filter( table_id_param INT, filter_param VARCHAR(1000) )
+CREATE PROCEDURE get_uploadTable_with_filter( table_id_param INT, user_id_param INT, filter_param VARCHAR(1000))
+-- if the filter_param contains any of these we additionaly filter: LAST_HOUR, LAST_DAY, LAST_WEEK, LAST_MONTH, LAST_YEAR
 BEGIN
     SET @@group_concat_max_len = 32000;
     SET @sql = NULL;
@@ -16,26 +17,44 @@ BEGIN
         )
     ) INTO @sql
     FROM  uploadTables t
-	INNER JOIN uploadRows r on (t.id = r.id_uploadTables)
-	INNER JOIN uploadCells cell on (cell.id_uploadRows = r.id)
-	INNER JOIN uploadCols col on (col.id = cell.id_uploadCols)
-    WHERE t.id = table_id_param;
+	INNER JOIN uploadCols col on (t.id = col.id_uploadTables)
+    WHERE t.id = table_id_param AND col.`name` != 'id_users';
 
     IF (@sql is null) THEN
         SELECT table_name from view_uploadTables where 1=2;
     ELSE
         BEGIN
-            SET @sql = CONCAT('select * from (select t.name as table_name, t.timestamp as timestamp, r.id as record_id, r.timestamp as entry_date, ', IF(@sql LIKE '%id_users%', @sql, CONCAT(@sql,', -1 AS id_users')), 
-                ' from uploadTables t
+			SET @user_filter = '';
+            IF user_id_param > 0 THEN
+				SET @user_filter = CONCAT(' AND r.id_users = ', user_id_param);
+            END IF;	
+            
+            SET @time_period_filter = '';
+            CASE 
+				WHEN filter_param LIKE '%LAST_HOUR%' THEN
+					SET @time_period_filter = ' AND r.`timestamp` >= CURDATE() - INTERVAL 1 HOUR';
+				WHEN filter_param LIKE '%LAST_DAY%' THEN
+					SET @time_period_filter = ' AND r.`timestamp` >= CURDATE() - INTERVAL 1 DAY';
+				WHEN filter_param LIKE '%LAST_WEEK%' THEN
+					SET @time_period_filter = ' AND r.`timestamp` >= CURDATE() - INTERVAL 1 WEEK';
+				WHEN filter_param LIKE '%LAST_MONTH%' THEN
+					SET @time_period_filter = ' AND r.`timestamp` >= CURDATE() - INTERVAL 1 MONTH';
+				WHEN filter_param LIKE '%LAST_YEAR%' THEN
+					SET @time_period_filter = ' AND r.`timestamp` >= CURDATE() - INTERVAL 1 YEAR';
+				ELSE
+					SET @time_period_filter = '';					
+			END CASE;
+            
+            SET @sql = CONCAT('select * from (select t.name as table_name, t.timestamp as timestamp, r.id as record_id, 
+					r.timestamp as entry_date, r.id_users, u.name as user_name,', @sql, 
+					' from uploadTables t
 					inner join uploadRows r on (t.id = r.id_uploadTables)
 					inner join uploadCells cell on (cell.id_uploadRows = r.id)
 					inner join uploadCols col on (col.id = cell.id_uploadCols)
-					where t.id = ', table_id_param,
+                    left join users u on (r.id_users = u.id)
+					where t.id = ', table_id_param, @user_filter, @time_period_filter,
 					' group by t.name, t.timestamp, r.id ) as r where 1=1  ', filter_param);
-			IF LOCATE('id_users', @sql) THEN
-				-- get user_name if there is id_users column
-				SET @sql = CONCAT('select v.*, u.name as user_name from (', @sql, ')  as v left join users u on (v.id_users = u.id) where 1=1  ', filter_param);
-			END IF;			
+            -- select @sql;
             PREPARE stmt FROM @sql;
             EXECUTE stmt;
             DEALLOCATE PREPARE stmt;