+++ server/db/FUN_PRO_VIEWS/16_proc_get_uploadTable.sql
@@ -4,7 +4,7 @@ DROP PROCEDURE IF EXISTS get_uploadTable //
 
 CREATE PROCEDURE get_uploadTable( table_id_param INT )
 BEGIN
-    CALL get_uploadTable_with_filter(table_id_param, '');
+    CALL get_uploadTable_with_filter(table_id_param, -1, '');
 END
 //